pyproject.toml
[project]
name = "auto-science"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.5.2",
    "dotenv>=0.9.9",
    "fpdf>=1.7.2",
    "google>=3.0.0",
    "google-api-core>=2.24.2",
    "google-genai>=1.8.0",
    "matplotlib>=3.10.1",
    "pandas>=2.2.3",
    "requests>=2.32.3",
]
